--- app/src/main/java/com/example/playlistmaker/data/TrackRepositoryImpl.kt
@@ -1,19 +1,0 @@
-package com.example.playlistmaker.data
-
-import com.example.playlistmaker.data.dto.TrackResponse
-import com.example.playlistmaker.data.dto.TrackSearchRequest
-import com.example.playlistmaker.domain.Track
-import com.example.playlistmaker.domain.api.TrackRepository
-
-class TrackRepositoryImpl(private val networkClient: NetworkClient) : TrackRepository {
-
-    override fun searchTracks(expression: String): List<Track> {
-        val response = networkClient.doRequest(TrackSearchRequest(expression))
-        if (response.resultCode == 200) {
-            return (response as TrackResponse).results.map {
-                Track(it.trackName, it.artistName, it.trackTimeMillis, it.artworkUrl100, it.trackId,it.collectionName,it.releaseDate,it.primaryGenreName,it.country,it.previewUrl) }
-        } else {
-            return emptyList()
-        }
-    }
-}
