+++ app/src/main/java/com/example/playlistmaker/data/network/RetrofitNetworkClient.kt
@@ -1,12 +1,15 @@
 package com.example.playlistmaker.data.network
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkCapabilities
 import com.example.playlistmaker.data.NetworkClient
 import com.example.playlistmaker.data.dto.TrackSearchRequest
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
-class RetrofitNetworkClient : NetworkClient {
+class RetrofitNetworkClient(private val context: Context) : NetworkClient {
 
     private val baseUrl = "https://itunes.apple.com"
 
@@ -18,14 +21,34 @@ class RetrofitNetworkClient : NetworkClient {
     private val itunesService = retrofit.create(ItunesAPI::class.java)
 
     override fun doRequest(dto: Any):com.example.playlistmaker.data.dto.Response  {
-        if (dto is TrackSearchRequest) {
-            val resp = itunesService.search(dto.expression).execute()
-
-            val body = resp.body() ?: com.example.playlistmaker.data.dto.Response()
-
-            return body.apply { resultCode = resp.code() }
-        } else {
+        if (isConnected() == false) {
+            return com.example.playlistmaker.data.dto.Response().apply { resultCode = -1 }
+        }
+        if (dto !is TrackSearchRequest) {
             return com.example.playlistmaker.data.dto.Response().apply { resultCode = 400 }
         }
+
+        val response = itunesService.search(dto.expression).execute()
+        val body = response.body()
+        return if (body != null) {
+            body.apply { resultCode = response.code() }
+        } else {
+            com.example.playlistmaker.data.dto.Response().apply { resultCode = response.code() }
+        }
     }
+
+    private fun isConnected(): Boolean {
+        val connectivityManager = context.getSystemService(
+            Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+        val capabilities = connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
+        if (capabilities != null) {
+            when {
+                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) -> return true
+                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI) -> return true
+                capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET) -> return true
+            }
+        }
+        return false
+    }
+
 }
