+++ app/src/main/java/com/example/playlistmaker/data/repository/TrackRepositoryImpl.kt
@@ -1,0 +1,41 @@
+package com.example.playlistmaker.data.repository
+
+import com.example.playlistmaker.data.NetworkClient
+import com.example.playlistmaker.data.dto.TrackResponse
+import com.example.playlistmaker.data.dto.TrackSearchRequest
+import com.example.playlistmaker.domain.Track
+import com.example.playlistmaker.domain.api.TrackRepository
+import com.example.playlistmaker.util.Resource
+
+class TrackRepositoryImpl(private val networkClient: NetworkClient) : TrackRepository {
+
+    override fun searchTracks(expression: String): Resource<List<Track>> {
+        val response = networkClient.doRequest(TrackSearchRequest(expression))
+        return when (response.resultCode) {
+            -1 -> {
+                Resource.Error("Проверьте подключение к интернету")
+            }
+
+            200 -> {
+                Resource.Success((response as TrackResponse).results.map {
+                    Track(
+                        it.trackName,
+                        it.artistName,
+                        it.trackTimeMillis,
+                        it.artworkUrl100,
+                        it.trackId,
+                        it.collectionName,
+                        it.releaseDate,
+                        it.primaryGenreName,
+                        it.country,
+                        it.previewUrl
+                    )
+                })
+            }
+
+            else -> {
+                Resource.Error("Ошибка сервера")
+            }
+        }
+    }
+}
