+++ app/src/main/java/com/example/playlistmaker/domain/api/TrackInteractor.kt
@@ -6,6 +6,6 @@ interface TrackInteractor {
     fun searchTracks(expression: String, consumer: TracksConsumer)
 
     interface TracksConsumer {
-        fun consume(foundTracks: List<Track>)
+        fun consume(foundTracks: List<Track>?,errorMessage: String?)
     }
 }
