+++ app/src/main/java/com/example/playlistmaker/domain/api/TrackRepository.kt
@@ -1,7 +1,8 @@
 package com.example.playlistmaker.domain.api
 
 import com.example.playlistmaker.domain.Track
+import com.example.playlistmaker.util.Resource
 
 interface TrackRepository {
-    fun searchTracks(expression: String): List<Track>
+    fun searchTracks(expression: String): Resource<List<Track>>
 }
