+++ app/src/main/java/com/example/playlistmaker/domain/impl/TrackInteractorImpl.kt
@@ -2,6 +2,7 @@ package com.example.playlistmaker.domain.impl
 
 import com.example.playlistmaker.domain.api.TrackInteractor
 import com.example.playlistmaker.domain.api.TrackRepository
+import com.example.playlistmaker.util.Resource
 import java.util.concurrent.Executors
 
 class TrackInteractorImpl (private val repository: TrackRepository) : TrackInteractor {
@@ -10,7 +11,10 @@ class TrackInteractorImpl (private val repository: TrackRepository) : TrackInter
 
     override fun searchTracks(expression: String, consumer: TrackInteractor.TracksConsumer) {
         executor.execute {
-            consumer.consume(repository.searchTracks(expression))
+            when(val resource = repository.searchTracks(expression)) {
+                is Resource.Success -> { consumer.consume(resource.data, null) }
+                is Resource.Error -> { consumer.consume(null, resource.message) }
+            }
         }
     }
 }
