+++ app/src/main/java/com/example/playlistmaker/presentation/TrackSearchController.kt
@@ -1,9 +1,7 @@
 package com.example.playlistmaker.presentation
 
-import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.ContentValues
-import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -19,20 +17,19 @@ import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.ProgressBar
 import android.widget.TextView
+import androidx.core.content.ContextCompat.getSystemService
 import androidx.recyclerview.widget.RecyclerView
-import com.example.playlistmaker.Creator
 import com.example.playlistmaker.R
 import com.example.playlistmaker.domain.Track
 import com.example.playlistmaker.domain.api.TrackInteractor
 import com.example.playlistmaker.ui.Adapter
-import com.example.playlistmaker.ui.PlayerActivity
+import com.example.playlistmaker.util.Creator
 
-class SearchController(private val activity: Activity,
-                       private val adapter: Adapter
+class TrackSearchController(
+    private val activity: Activity,
+    private val adapter: Adapter,
+    private val adapterHistory: Adapter
 ) {
-
-
-
     companion object {
         const val SAVED_INPUT = "SAVED_INPUT"
         private const val CLICK_DEBOUNCE_DELAY = 1000L
@@ -42,6 +39,10 @@ class SearchController(private val activity: Activity,
     private var isClickAllowed = true
 
     private val handler = Handler(Looper.getMainLooper())
+
+    private val tracksInteractor = Creator.provideTrackInteractor(activity)
+    private val trackList = arrayListOf<Track>()
+
 
     private lateinit var inputText: String
     private lateinit var input: EditText
@@ -57,13 +58,14 @@ class SearchController(private val activity: Activity,
     private lateinit var historyRecycler: RecyclerView
     private lateinit var progressBar: ProgressBar
 
-    private val trackList = arrayListOf<Track>()
-    private val adapterHistory = adapter
-    private  val tracksInteractor = Creator.provideTrackInteractor()
+    private val searchRunnable = Runnable { searchRequest(inputEditText) }
+    private fun searchDebounce() {
+        handler.removeCallbacks(searchRunnable)
+        handler.postDelayed(searchRunnable, SEARCH_DEBOUNCE_DELAY)
+    }
 
-    @SuppressLint("MissingInflatedId")
-   fun onCreate(savedInstanceState: Bundle?) {
-
+    fun onCreate() {
+        //
         backButton = activity.findViewById(R.id.back)
         inputEditText = activity.findViewById(R.id.inputEditText)
         clearButton = activity.findViewById(R.id.clearIcon)
@@ -79,43 +81,34 @@ class SearchController(private val activity: Activity,
         inputText = ""
         input = inputEditText
 
-        //var sharedPreferences = getSharedPreferences(TRACK_SEARCH_HISTORY, MODE_PRIVATE)
-        //val history = SearchHistory(sharedPreferences).read()?.toCollection(ArrayList())
-        val searchHistoryProvider = Creator.provideSearchHistory(activity.applicationContext)
-        val history = searchHistoryProvider.read()
+        val adapterHistory = adapterHistory
 
         adapter.tracks = trackList
         recyclerView.adapter = adapter
 
+        val searchHistoryProvider = Creator.provideSearchHistory(activity.applicationContext)
+        val history = searchHistoryProvider.read()
 
         if (history != null && history.size != 0) {
-            //sharedPreferences = getSharedPreferences(TRACK_SEARCH_HISTORY, MODE_PRIVATE)
-            adapterHistory.tracks =
-                    //SearchHistory(sharedPreferences).read()?.toCollection(ArrayList())!!
-                searchHistoryProvider.read()?.toCollection(ArrayList())!!
+
+            adapterHistory.tracks = searchHistoryProvider.read()?.toCollection(ArrayList())!!
             historyRecycler.adapter = adapterHistory
             historyView.visibility = View.VISIBLE
 
         }
 
-
+        //Buttons actions////////////////////////////////////////////////////////////
         clearHistoryButton.setOnClickListener {
             Log.v(ContentValues.TAG, "history is  $history");
-            //SearchHistory(sharedPreferences).clear()
             searchHistoryProvider.clear()
             historyView.visibility = View.GONE
 
         }
+
         backButton.setOnClickListener {
             activity.finish()
         }
-        inputEditText.setOnEditorActionListener { _, actionId, _ ->
-            if (actionId == EditorInfo.IME_ACTION_DONE) {
-                searchRequest(inputEditText)//search(inputEditText)
-                true
-            }
-            false
-        }
+
         placeholderButton.setOnClickListener { searchRequest(input) }
 
         clearButton.setOnClickListener {
@@ -143,13 +136,24 @@ class SearchController(private val activity: Activity,
             val view: View? = activity.currentFocus
             val inputMethodManager =
                 activity.getSystemService(Activity.INPUT_METHOD_SERVICE) as InputMethodManager
-
             // on below line hiding our keyboard.
             if (view != null) {
                 inputMethodManager.hideSoftInputFromWindow(view.windowToken, 0)
             }
 
         }
+
+        /////////////////////////////////////////////////////////////////////////////////
+
+        //EditText actions
+        inputEditText.setOnEditorActionListener { _, actionId, _ ->
+            if (actionId == EditorInfo.IME_ACTION_DONE) {
+                searchRequest(inputEditText)//search(inputEditText)
+                true
+            }
+            false
+        }
+
         val simpleTextWatcher = object : TextWatcher {
             override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
                 // empty
@@ -171,22 +175,42 @@ class SearchController(private val activity: Activity,
         inputEditText.addTextChangedListener(simpleTextWatcher)
 
 
+        //////////////////////////////////////////////////////////////////////////////////////
+
+
     }
 
-    private val searchRunnable = Runnable { searchRequest(inputEditText) }
+    ////Functions
 
-    private fun searchDebounce() {
-        handler.removeCallbacks(searchRunnable)
-        handler.postDelayed(searchRunnable, SEARCH_DEBOUNCE_DELAY)
-    }
+    private fun searchRequest(textToSearch: EditText) {
+        if (textToSearch.text.isNotEmpty()) {
 
-    private fun clickDebounce(): Boolean {
-        val current = isClickAllowed
-        if (isClickAllowed) {
-            isClickAllowed = false
-            handler.postDelayed({ isClickAllowed = true }, CLICK_DEBOUNCE_DELAY)
+            progressBar.visibility = View.VISIBLE
+            placeholderMessage.visibility = View.GONE
+            recyclerView.visibility = View.GONE
+
+            tracksInteractor.searchTracks(
+                textToSearch.text.toString(),
+                object : TrackInteractor.TracksConsumer {
+                    override fun consume(foundTracks: List<Track>?, errorMessage: String?) {
+                        handler.post {
+                            if (foundTracks != null) {
+                                progressBar.visibility = View.GONE
+                                trackList.clear()
+                                trackList.addAll(foundTracks)
+                                recyclerView.visibility = View.VISIBLE
+                                adapter.notifyDataSetChanged()
+                            }
+                            if (errorMessage != null) {
+                                showMessage(false)
+                            } else if (trackList.isEmpty()) {
+                                //Log.v(ContentValues.TAG, "Something here")
+                                showMessage(true)
+                            }
+                        }
+                    }
+                })
         }
-        return current
     }
 
     private fun showMessage(noDataFound: Boolean) {
@@ -214,50 +238,6 @@ class SearchController(private val activity: Activity,
         }
     }
 
-    private fun putExtra(intent: Intent, track: Track) {
-        intent.putExtra("trackName", track.trackName)
-        intent.putExtra("artistName", track.artistName)
-        intent.putExtra("trackTimeMillis", track.trackTimeMillis)
-        intent.putExtra("artworkUrl100", track.artworkUrl100)
-        intent.putExtra("collectionName", track.collectionName)
-        intent.putExtra("releaseDate", track.releaseDate)
-        intent.putExtra("primaryGenreName", track.primaryGenreName)
-        intent.putExtra("country", track.country)
-        intent.putExtra("previewUrl", track.previewUrl)
-    }
-
-
-
-    private fun searchRequest(textToSearch: EditText) {
-        if (textToSearch.text.isNotEmpty()) {
-
-            progressBar.visibility = View.VISIBLE
-            placeholderMessage.visibility = View.GONE
-            recyclerView.visibility = View.GONE
-
-            tracksInteractor.searchTracks(textToSearch.text.toString(), object : TrackInteractor.TracksConsumer {
-                override fun consume(foundTracks: List<Track>) {
-                    handler.post {
-                        if(foundTracks.isNotEmpty()){
-                            progressBar.visibility = View.GONE
-                            trackList.clear()
-                            trackList.addAll(foundTracks)
-                            recyclerView.visibility = View.VISIBLE
-                            adapter.notifyDataSetChanged()
-                        }
-                        if (foundTracks.isEmpty()) {
-                            showMessage(true)
-                        } //else {
-                        // Log.v(ContentValues.TAG, "Something here")
-                        // showMessage(false)
-                        //}
-                    }
-                }
-            })
-        }
-    }
-
-
     private fun clearButtonVisibility(s: CharSequence?): Int {
         return if (s.isNullOrEmpty()) {
             View.GONE
@@ -265,5 +245,20 @@ class SearchController(private val activity: Activity,
             View.VISIBLE
         }
     }
-}
+
+    fun onSaveInstanceState(outState: Bundle) {
+        outState.putString(SAVED_INPUT, inputText)
+    }
+
+    fun onRestoreInstanceState(savedInstanceState: Bundle) {
+        inputText = savedInstanceState.getString(SAVED_INPUT).toString()
+        input.setText(inputText)
+
+    }
+
+    fun onDestroy() {
+        handler.removeCallbacks(searchRunnable)
+    }
+
+
 }
