+++ app/src/main/java/com/example/playlistmaker/util/Resource.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.util
+
+sealed class Resource<T>(val data: T? = null, val message: String? = null) {
+    class Success<T>(data: T): Resource<T>(data)
+    class Error<T>(message: String, data: T? = null): Resource<T>(data, message)
+}
